Add ClassfulLinkableMixin
