+++ packages/template_components/mixins/__init__.py
@@ -1,0 +1,3 @@
+from .classful_linkable import ClassfulLinkableMixin
+
+__all__ = ["ClassfulLinkableMixin"]
